cleanup obsolete tests and rename a file
