+++ server/controllers/items/lib/validate_entity_and_shelves.js
